mesa update until Sep 18 2021
mesa             commit 6b5e9352ef7661670d9873e0cf96d6d5d2174c6f

+++ mesalib/.gitlab-ci/piglit/run.sh
@@ -3,6 +3,7 @@
 set -ex
 
 INSTALL=$(realpath -s "$PWD"/install)
+MINIO_ARGS="--credentials=/tmp/.minio_credentials"
 
 RESULTS=$(realpath -s "$PWD"/results)
 mkdir -p "$RESULTS"
@@ -15,7 +16,22 @@ export __LD_LIBRARY_PATH="$LD_LIBRARY_PATH:$INSTALL/lib/"
 
 # Sanity check to ensure that our environment is sufficient to make our tests
 # run against the Mesa built by CI, rather than any installed distro version.
-MESA_VERSION=$(cat "$INSTALL/VERSION" | sed 's/\./\\./g')
+MESA_VERSION=$(head -1 "$INSTALL/VERSION" | sed 's/\./\\./g')
+
+print_red() {
+    RED='\033[0;31m'
+    NC='\033[0m' # No Color
+    printf "${RED}"
+    "$@"
+    printf "${NC}"
+}
+
+# wrapper to supress +x to avoid spamming the log
+quiet() {
+    set +x
+    "$@"
+    set -x
+}
 
 if [ "$VK_DRIVER" ]; then
 
@@ -40,18 +56,23 @@ if [ "$VK_DRIVER" ]; then
 
     SANITY_MESA_VERSION_CMD="vulkaninfo"
 
+    HANG_DETECTION_CMD="/parallel-deqp-runner/build/bin/hang-detection"
+
 
     # Set up the Window System Interface (WSI)
 
-    # IMPORTANT:
-    #
-    # Nothing to do here.
-    #
-    # Run vulkan against the host's running X server (xvfb doesn't
-    # have DRI3 support).
-    # Set the DISPLAY env variable in each gitlab-runner's
-    # configuration file:
-    # https://docs.gitlab.com/runner/configuration/advanced-configuration.html#the-runners-section
+    if [ ${TEST_START_XORG:-0} -eq 1 ]; then
+        "$INSTALL"/common/start-x.sh "$INSTALL"
+        export DISPLAY=:0
+    else
+        # Run vulkan against the host's running X server (xvfb doesn't
+        # have DRI3 support).
+        # Set the DISPLAY env variable in each gitlab-runner's
+        # configuration file:
+        # https://docs.gitlab.com/runner/configuration/advanced-configuration.html#the-runners-section
+        quiet printf "%s%s\n" "Running against the hosts' X server. " \
+              "DISPLAY is \"$DISPLAY\"."
+    fi
 else
 
     ### GL/ES ###
@@ -66,6 +87,8 @@ else
     fi
 
     SANITY_MESA_VERSION_CMD="wflinfo"
+
+    HANG_DETECTION_CMD=""
 
 
     # Set up the platform windowing system.
@@ -84,7 +107,7 @@ else
 
             LD_LIBRARY_PATH="$__LD_LIBRARY_PATH" \
             GALLIUM_DRIVER=llvmpipe \
-            GALLIVM_PERF="nopt,no_filter_hacks" \
+            GALLIVM_PERF="nopt" \
             VTEST_USE_EGL_SURFACELESS=1 \
             VTEST_USE_GLES=1 \
             virgl_test_server >"$RESULTS"/vtest-log.txt 2>&1 &
@@ -120,21 +143,6 @@ if [ -n "$CI_NODE_INDEX" ]; then
     USE_CASELIST=1
 fi
 
-print_red() {
-    RED='\033[0;31m'
-    NC='\033[0m' # No Color
-    printf "${RED}"
-    "$@"
-    printf "${NC}"
-}
-
-# wrapper to supress +x to avoid spamming the log
-quiet() {
-    set +x
-    "$@"
-    set -x
-}
-
 replay_minio_upload_images() {
     find "$RESULTS/$__PREFIX" -type f -name "*.png" -printf "%P\n" \
         | while read -r line; do
@@ -144,36 +152,26 @@ replay_minio_upload_images() {
             if [ "x$CI_PROJECT_PATH" != "x$FDO_UPSTREAM_REPO" ]; then
                 continue
             fi
-            __MINIO_PATH="$PIGLIT_REPLAY_REFERENCE_IMAGES_BASE_URL"
+            __MINIO_PATH="$PIGLIT_REPLAY_REFERENCE_IMAGES_BASE"
             __DESTINATION_FILE_PATH="${line##*-}"
-            if ci-fairy minio ls "minio://${MINIO_HOST}${__MINIO_PATH}/${__DESTINATION_FILE_PATH}" 2>/dev/null; then
+            if wget -q --method=HEAD "https://${__MINIO_PATH}/${__DESTINATION_FILE_PATH}" 2>/dev/null; then
                 continue
             fi
         else
-            __MINIO_PATH="$PIGLIT_REPLAY_ARTIFACTS_BASE_URL"
+            __MINIO_PATH="$JOB_ARTIFACTS_BASE"
             __DESTINATION_FILE_PATH="$__MINIO_TRACES_PREFIX/${line##*-}"
-            # Adding to the JUnit the direct link to the diff page in
-            # the dashboard
-            __PIGLIT_TESTCASE_CLASSNAME="piglit\.trace\.$PIGLIT_REPLAY_DEVICE_NAME\.$(dirname $__TRACE | sed 's%/%\\.%g;s@%@\\%@')"
-            __PIGLIT_TESTCASE_NAME="$(basename $__TRACE | sed 's%\.%_%g;s@%@\\%@')"
-            __DASHBOARD_URL="https://tracie.freedesktop.org/dashboard/imagediff/${CI_PROJECT_PATH}/${CI_JOB_ID}/${__TRACE}"
-            __START_TEST_PATTERN='<testcase classname="'"${__PIGLIT_TESTCASE_CLASSNAME}"'" name="'"${__PIGLIT_TESTCASE_NAME}"'" status="fail"'
-            __REPLACE_TEST_PATTERN='</system-out><failure type="fail"/></testcase>'
-            # Replace in the range between __START_TEST_PATTERN and
-            # __REPLACE_TEST_PATTERN leaving __START_TEST_PATTERN out
-            # from the substitution
-            sed '\%'"${__START_TEST_PATTERN}"'%,\%'"${__REPLACE_TEST_PATTERN}"'%{\%'"${__START_TEST_PATTERN}"'%b;s%'"${__REPLACE_TEST_PATTERN}"'%</system-out><failure type="fail">To view the image differences visit: '"${__DASHBOARD_URL}"'</failure></testcase>%}' \
-                -i "$RESULTS"/junit.xml
         fi
 
-        ci-fairy minio cp "$RESULTS/$__PREFIX/$line" \
-            "minio://${MINIO_HOST}${__MINIO_PATH}/${__DESTINATION_FILE_PATH}"
+        ci-fairy minio cp $MINIO_ARGS "$RESULTS/$__PREFIX/$line" \
+            "minio://${__MINIO_PATH}/${__DESTINATION_FILE_PATH}"
     done
 }
 
 SANITY_MESA_VERSION_CMD="$SANITY_MESA_VERSION_CMD | tee /tmp/version.txt | grep \"Mesa $MESA_VERSION\(\s\|$\)\""
 
-rm -rf results
+if [ -d results ]; then
+    cd results && rm -rf ..?* .[!.]* *
+fi
 cd /piglit
 
 if [ -n "$USE_CASELIST" ]; then
@@ -194,13 +192,18 @@ PIGLIT_TESTS=$(printf "%s" "$PIGLIT_TESTS")
 
 PIGLIT_CMD="./piglit run --timeout 300 -j${FDO_CI_CONCURRENT:-4} $PIGLIT_OPTIONS $PIGLIT_TESTS $PIGLIT_PROFILES "$(/usr/bin/printf "%q" "$RESULTS")
 
-RUN_CMD="export LD_LIBRARY_PATH=$__LD_LIBRARY_PATH; $SANITY_MESA_VERSION_CMD && $PIGLIT_CMD"
+RUN_CMD="export LD_LIBRARY_PATH=$__LD_LIBRARY_PATH; $SANITY_MESA_VERSION_CMD && $HANG_DETECTION_CMD $PIGLIT_CMD"
 
 if [ "$RUN_CMD_WRAPPER" ]; then
     RUN_CMD="set +e; $RUN_CMD_WRAPPER "$(/usr/bin/printf "%q" "$RUN_CMD")"; set -e"
 fi
 
 FAILURE_MESSAGE=$(printf "%s" "Unexpected change in results:")
+
+if [ "x$PIGLIT_PROFILES" = "xreplay" ] \
+       && [ ${PIGLIT_REPLAY_UPLOAD_TO_MINIO:-0} -eq 1 ]; then
+    ci-fairy minio login $MINIO_ARGS $CI_JOB_JWT
+fi
 
 eval $RUN_CMD
 
@@ -227,13 +230,13 @@ mkdir -p .gitlab-ci/piglit
 if [ "x$PIGLIT_PROFILES" = "xreplay" ] \
        && [ ${PIGLIT_REPLAY_UPLOAD_TO_MINIO:-0} -eq 1 ]; then
 
-    ci-fairy minio login $CI_JOB_JWT
-
     __PREFIX="trace/$PIGLIT_REPLAY_DEVICE_NAME"
     __MINIO_PATH="$PIGLIT_REPLAY_ARTIFACTS_BASE_URL"
     __MINIO_TRACES_PREFIX="traces"
 
-    quiet replay_minio_upload_images
+    if [ "x$PIGLIT_REPLAY_SUBCOMMAND" != "xprofile" ]; then
+        quiet replay_minio_upload_images
+    fi
 fi
 
 if [ -n "$USE_CASELIST" ]; then
@@ -244,9 +247,11 @@ if [ -n "$USE_CASELIST" ]; then
 
     grep -F -f /tmp/executed.txt "$INSTALL/$PIGLIT_RESULTS.txt" \
        > ".gitlab-ci/piglit/$PIGLIT_RESULTS.txt.baseline" || true
-else
+elif [ -f "$INSTALL/$PIGLIT_RESULTS.txt" ]; then
     cp "$INSTALL/$PIGLIT_RESULTS.txt" \
        ".gitlab-ci/piglit/$PIGLIT_RESULTS.txt.baseline"
+else
+    touch ".gitlab-ci/piglit/$PIGLIT_RESULTS.txt.baseline"
 fi
 
 if diff -q ".gitlab-ci/piglit/$PIGLIT_RESULTS.txt.baseline" $RESULTSFILE; then
@@ -258,9 +263,9 @@ fi
 
 if [ "x$PIGLIT_PROFILES" = "xreplay" ]; then
 find "$RESULTS"/summary -type f -name "*.html" -print0 \
-        | xargs -0 sed -i 's%<img src="file://'"${RESULTS}"'.*-\([0-9a-f]*\)\.png%<img src="https://'"${MINIO_HOST}${PIGLIT_REPLAY_ARTIFACTS_BASE_URL}"'/traces/\1.png%g'
+        | xargs -0 sed -i 's%<img src="file://'"${RESULTS}"'.*-\([0-9a-f]*\)\.png%<img src="https://'"${JOB_ARTIFACTS_BASE}"'/traces/\1.png%g'
 find "$RESULTS"/summary -type f -name "*.html" -print0 \
-        | xargs -0 sed -i 's%<img src="file://%<img src="https://'"${MINIO_HOST}${PIGLIT_REPLAY_REFERENCE_IMAGES_BASE_URL}"'/%g'
+        | xargs -0 sed -i 's%<img src="file://%<img src="https://'"${PIGLIT_REPLAY_REFERENCE_IMAGES_BASE}"'/%g'
 fi
 
 FAILURE_MESSAGE=$(printf "${FAILURE_MESSAGE}\n%s" "Check the HTML summary for problems at: ${ARTIFACTS_BASE_URL}/results/summary/problems.html")
